[refactor/project_structure]: 커스텀 뷰 패키지 변경
