+++ presentation/src/main/java/org/gdsc/presentation/view/custom/JmtEditText.kt
@@ -1,4 +1,4 @@
-package org.gdsc.presentation.view
+package org.gdsc.presentation.view.custom
 
 import androidx.constraintlayout.widget.ConstraintLayout
 import android.content.Context
