+++ presentation/src/main/java/org/gdsc/presentation/view/custom/JmtSnackbar.kt
@@ -1,4 +1,4 @@
-package org.gdsc.presentation.view
+package org.gdsc.presentation.view.custom
 
 import android.view.LayoutInflater
 import android.view.View
